src/requirements.txt
asgiref==3.5.0
pypi
psycopg2-binary==2.9.3
django-redis==4.12.1
Django==4.0.4
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2021.5
celery==5.2.3
djangorestframework==3.13.1
python-slugify==6.1.1
djangorestframework-simplejwt
pytest
pytest-django
flake8